# backend/swagger.yaml
openapi: 3.0.3
info:
  title: SaaS Multi-Tenant Backend Platform API
  description: |
    A comprehensive multi-tenant SaaS backend platform built with Flask.

    ## Features
    - Multi-tenant architecture with database isolation
    - JWT-based authentication with refresh tokens
    - File management with MD5 deduplication
    - Role-based access control (Admin, User, Viewer)
    - Document management system
    - Kafka integration (placeholder for Phase 6)

    ## Authentication
    Most endpoints require JWT authentication. Include the access token in the Authorization header:
    ```
    Authorization: Bearer <access_token>
    ```

    ## Rate Limiting
    API rate limiting will be implemented in Phase 6.

    ## Error Responses
    All endpoints return consistent error responses with the following structure:
    ```json
    {
      "success": false,
      "message": "Error description",
      "error": "ERROR_CODE"
    }
    ```
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:4999
    description: Development server
  - url: http://localhost:8080
    description: Docker development server
  - url: https://api.example.com
    description: Production server (Phase 7)
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token
  schemas:
    Error:
      type: object
      required:
        - success
        - message
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: "An error occurred"
        error:
          type: string
          example: "ERROR_CODE"
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        first_name:
          type: string
          example: "John"
        last_name:
          type: string
          example: "Doe"
        email:
          type: string
          format: email
          example: "john.doe@example.com"
        is_active:
          type: boolean
          example: true
        created_at:
          type: string
          format: date-time
          example: "2024-01-01T00:00:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2024-01-01T00:00:00Z"
...
    SSOAvailability:
      type: object
      properties:
        available:
          type: boolean
          example: true
        auth_method:
          type: string
          enum: [local, sso, both]
          example: "both"
        provider:
          type: string
          enum: [azure_ad, null]
          example: "azure_ad"
        auto_provisioning:
          type: boolean
          example: true
        login_url:
          type: string
          format: uri
          example: "/api/auth/sso/azure/login/123e4567-e89b-12d3-a456-426614174000"
    SSOLoginResponse:
      type: object
      properties:
        access_token:
          type: string
          example: "jwt-access-token"
        refresh_token:
          type: string
          example: "jwt-refresh-token"
        user:
          $ref: '#/components/schemas/User'
        tenant_id:
          type: string
          format: uuid
        auth_method:
          type: string
          enum: [sso, local]
          example: "sso"
        azure_identity:
          type: object
          properties:
            object_id:
              type: string
            upn:
              type: string
            display_name:
              type: string
paths:
...
  /api/auth/sso/check-availability/{tenant_id}:
    get:
      tags:
        - SSO
      summary: Check SSO availability
      description: |
        Checks if SSO is available and configured for a specific tenant.
        Returns the authentication method and SSO provider details.
      parameters:
        - name: tenant_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: UUID of the tenant
      responses:
        '200':
          description: SSO availability status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SSOAvailability'
        '404':
          description: Tenant not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/auth/sso/azure/login/{tenant_id}:
    get:
      tags:
        - SSO
      summary: Initiate Azure AD login
      description: |
        Initiates the Azure AD OAuth2 authorization flow with PKCE.
        Redirects the user to Microsoft login page.

        **Flow:**
        1. Generates PKCE code challenge and state token
        2. Stores PKCE verifier and state in session
        3. Redirects to Microsoft authorization endpoint
        4. Microsoft redirects back to callback after authentication
      parameters:
        - name: tenant_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: UUID of the tenant
        - name: redirect_uri
          in: query
          schema:
            type: string
            format: uri
          description: Optional custom redirect URI after successful login
      responses:
        '302':
          description: Redirect to Azure AD login page
          headers:
            Location:
              description: Microsoft authorization URL
              schema:
                type: string
                format: uri
                example: "https://login.microsoftonline.com/{tenant}/oauth2/v2.0/authorize?..."
        '404':
          description: SSO not configured for tenant
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/auth/sso/azure/callback:
    get:
      tags:
        - SSO
      summary: Azure AD OAuth2 callback
      description: |
        Handles the OAuth2 callback from Azure AD after user authentication.
        Exchanges authorization code for tokens using PKCE.
        Creates or updates user account and returns JWT tokens.

        **Flow:**
        1. Validates state token (CSRF protection)
        2. Exchanges auth code for tokens using PKCE verifier
        3. Fetches user profile from Microsoft Graph
        4. Creates/updates user and Azure identity records
        5. Returns JWT tokens for the application
      parameters:
        - name: code
          in: query
          required: true
          schema:
            type: string
          description: Authorization code from Azure AD
        - name: state
          in: query
          required: true
          schema:
            type: string
          description: State token for CSRF protection
        - name: error
          in: query
          schema:
            type: string
          description: Error code if authentication failed
        - name: error_description
          in: query
          schema:
            type: string
          description: Error description if authentication failed
      responses:
        '200':
          description: SSO login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SSOLoginResponse'
        '400':
          description: Invalid state or authorization code
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Authentication failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/auth/sso/azure/refresh:
    post:
      tags:
        - SSO
      summary: Refresh Azure AD tokens
      description: |
        Refreshes Azure AD access token using stored refresh token.
        This is called automatically by background tasks before token expiry.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - tenant_id
              properties:
                tenant_id:
                  type: string
                  format: uuid
                  description: UUID of the tenant
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Azure AD token refreshed successfully"
                  expires_at:
                    type: string
                    format: date-time
        '401':
          description: Refresh token invalid or expired
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/auth/sso/azure/logout/{tenant_id}:
    post:
      tags:
        - SSO
      summary: Azure AD logout
      description: |
        Logs out the user from Azure AD SSO session.
        Clears stored tokens and optionally redirects to Microsoft logout.
      security:
        - BearerAuth: []
      parameters:
        - name: tenant_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: UUID of the tenant
        - name: global_logout
          in: query
          schema:
            type: boolean
            default: false
          description: Also logout from Microsoft account globally
      responses:
        '200':
          description: Logged out successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "SSO logout successful"
                  logout_url:
                    type: string
                    format: uri
                    description: Microsoft logout URL if global_logout=true
...